Add tests and refact ui.scheduler.js
