+++ testing/tests/DevExpress.ui.widgets.scheduler/common.tests.js
@@ -1,5 +1,4 @@
 import fx from 'animation/fx';
-import translator from 'animation/translator';
 import Color from 'color';
 import 'common.css!';
 import config from 'core/config';
@@ -10,7 +9,7 @@ import resizeCallbacks from 'core/utils/resize_callbacks';
 import { isRenderer } from 'core/utils/type';
 import CustomStore from 'data/custom_store';
 import { DataSource } from 'data/data_source/data_source';
-import dragEvents from 'events/drag';
+
 import { triggerHidingEvent, triggerShownEvent } from 'events/visibility_change';
 import 'generic_light.css!';
 import $ from 'jquery';
@@ -25,7 +24,6 @@ import keyboardMock from '../../helpers/keyboardMock.js';
 import pointerMock from '../../helpers/pointerMock.js';
 import { createWrapper, SchedulerTestWrapper } from '../../helpers/scheduler/helpers.js';
 import { Deferred } from 'core/utils/deferred';
-import timeZoneUtils from 'ui/scheduler/utils.timeZone';
 
 QUnit.testStart(function() {
     $('#qunit-fixture').html('<div id="scheduler"></div>');
@@ -2359,615 +2357,6 @@ QUnit.module('Scrolling to time', () => {
             this.clock.restore();
             fx.off = false;
         }
-    });
-
-    QUnit.test('onAppointmentAdding', function(assert) {
-        const addingSpy = sinon.spy(noop);
-
-        this.createInstance({
-            onAppointmentAdding: addingSpy,
-            dataSource: []
-        });
-
-        const newAppointment = {
-            startDate: new Date(2015, 1, 9, 16),
-            endDate: new Date(2015, 1, 9, 17),
-            text: 'caption'
-        };
-
-        this.instance.addAppointment(newAppointment);
-        this.clock.tick();
-
-
-        const args = addingSpy.getCall(0).args[0];
-
-        assert.ok(addingSpy.calledOnce, 'onAppointmentAdding was called');
-        assert.equal(args.element, this.instance.element(), 'Element field is OK');
-        assert.equal(args.component, this.instance, 'Component field is OK');
-        assert.strictEqual(args.cancel, false, '\'Cancel\' flag is OK');
-        assert.deepEqual(args.appointmentData, newAppointment, 'Appointment field is OK');
-    });
-
-    QUnit.test('Appointment should not be added to the data source if \'cancel\' flag is defined as true', function(assert) {
-        const dataSource = new DataSource({
-            store: []
-        });
-        this.createInstance({
-            onAppointmentAdding: function(args) {
-                args.cancel = true;
-            },
-            dataSource: dataSource
-        });
-
-        this.instance.addAppointment({ startDate: new Date(), text: 'Appointment 1' });
-        this.clock.tick();
-
-        assert.strictEqual(dataSource.items().length, 0, 'Insert operation is canceled');
-    });
-
-    QUnit.test('Appointment should not be added to the data source if \'cancel\' flag is defined as true during async operation', function(assert) {
-        const dataSource = new DataSource({
-            store: []
-        });
-        this.createInstance({
-            onAppointmentAdding: function(args) {
-                args.cancel = $.Deferred();
-                setTimeout(function() {
-                    args.cancel.resolve(true);
-                }, 200);
-            },
-            dataSource: dataSource
-        });
-
-        this.instance.addAppointment({ startDate: new Date(), text: 'Appointment 1' });
-        this.clock.tick(200);
-
-        assert.strictEqual(dataSource.items().length, 0, 'Insert operation is canceled');
-    });
-
-    QUnit.test('Appointment should not be added to the data source if \'cancel\' flag is defined as Promise', function(assert) {
-        const promise = new Promise(function(resolve) {
-            setTimeout(function() {
-                resolve(true);
-            }, 200);
-        });
-        const dataSource = new DataSource({
-            store: []
-        });
-        this.createInstance({
-            onAppointmentAdding: function(args) {
-                args.cancel = promise;
-            },
-            dataSource: dataSource
-        });
-
-        this.instance.addAppointment({ startDate: new Date(), text: 'Appointment 1' });
-        this.clock.tick(200);
-
-        promise.then(function() {
-            assert.strictEqual(dataSource.items().length, 0, 'Insert operation is canceled');
-        });
-
-        return promise;
-    });
-
-    QUnit.test('onAppointmentAdded', function(assert) {
-        const addedSpy = sinon.spy(noop);
-
-        this.createInstance({
-            onAppointmentAdded: addedSpy,
-            dataSource: []
-        });
-
-        const newAppointment = { startDate: new Date(2015, 1, 9, 16), endDate: new Date(2015, 1, 9, 17), text: 'caption' };
-
-        this.instance.addAppointment(newAppointment);
-        this.clock.tick();
-
-        const args = addedSpy.getCall(0).args[0];
-
-        assert.ok(addedSpy.calledOnce, 'onAppointmentAdded was called');
-        assert.deepEqual(args.appointmentData, newAppointment, 'Appointment field is OK');
-        assert.equal(args.element, this.instance.element(), 'Element field is OK');
-        assert.equal(args.component, this.instance, 'Component field is OK');
-        assert.strictEqual(args.error, undefined, 'Error field is not defined');
-    });
-
-    QUnit.test('onAppointmentAdded should have error field in args if an error occurs while data inserting', function(assert) {
-        const addedSpy = sinon.spy(noop);
-
-        this.createInstance({
-            onAppointmentAdded: addedSpy,
-            dataSource: new DataSource({
-                store: new CustomStore({
-                    load: noop,
-                    insert: function() {
-                        return $.Deferred().reject(new Error('Unknown error occurred'));
-                    }
-                })
-            })
-        });
-
-        this.instance.addAppointment({ startDate: new Date(2015, 1, 9, 16), endDate: new Date(2015, 1, 9, 17), text: 'caption' });
-        this.clock.tick();
-
-        const error = addedSpy.getCall(0).args[0].error;
-
-        assert.ok(error instanceof Error, 'Error field is defined');
-        assert.equal(error.message, 'Unknown error occurred', 'Error message is OK');
-    });
-
-
-    QUnit.test('onAppointmentUpdating', function(assert) {
-        const updatingSpy = sinon.spy(noop);
-        const oldData = [{ startDate: new Date(2015, 1, 9, 16), endDate: new Date(2015, 1, 9, 17), text: 'caption' }];
-        const newData = { startDate: new Date(2015, 1, 10, 16), endDate: new Date(2015, 1, 10, 17), text: 'title' };
-
-        this.createInstance({
-            onAppointmentUpdating: updatingSpy,
-            dataSource: new DataSource({ store: oldData })
-        });
-
-        this.instance.updateAppointment($.extend({}, oldData[0]), newData);
-        this.clock.tick();
-
-        const args = updatingSpy.getCall(0).args[0];
-
-        assert.ok(updatingSpy.calledOnce, 'onAppointmentUpdating was called');
-        assert.equal(args.element, this.instance.element(), 'Element field is OK');
-        assert.equal(args.component, this.instance, 'Component field is OK');
-        assert.strictEqual(args.cancel, false, '\'Cancel\' flag is OK');
-        assert.deepEqual(args.newData, newData, 'newData field is OK');
-        assert.deepEqual(args.oldData, oldData[0], 'oldData field is OK');
-    });
-
-    QUnit.test('Appointment should not be updated if \'cancel\' flag is defined as true', function(assert) {
-        const appointments = [{ startDate: new Date(2015, 1, 9, 16), endDate: new Date(2015, 1, 9, 17), text: 'caption' }];
-        const dataSource = new DataSource({
-            store: appointments
-        });
-
-        this.createInstance({
-            onAppointmentUpdating: function(args) {
-                args.cancel = true;
-            },
-            dataSource: dataSource,
-            currentDate: new Date(2015, 1, 9)
-        });
-
-        this.instance.updateAppointment(appointments[0], { startDate: new Date(), text: 'Appointment 1' });
-        this.clock.tick();
-
-        assert.deepEqual(dataSource.items(), [{ startDate: new Date(2015, 1, 9, 16), endDate: new Date(2015, 1, 9, 17), text: 'caption' }], 'Update operation is canceled');
-    });
-
-    QUnit.test('Appointment form should not be updated if \'cancel\' flag is defined as true (T653358)', function(assert) {
-        const tzOffsetStub = sinon.stub(timeZoneUtils, 'getClientTimezoneOffset').returns(-10800000);
-
-        try {
-            const appointments = [{ startDate: new Date(2015, 1, 9, 16), endDate: new Date(2015, 1, 9, 17), text: 'caption' }];
-            const dataSource = new DataSource({
-                store: appointments
-            });
-
-            this.createInstance({
-                timeZone: 'Etc/UTC',
-                onAppointmentUpdating: function(args) {
-                    args.cancel = true;
-                },
-                dataSource: dataSource,
-                currentDate: new Date(2015, 1, 9)
-            });
-
-            this.instance.showAppointmentPopup(appointments[0]);
-            $('.dx-scheduler-appointment-popup .dx-popup-done').trigger('dxclick');
-
-            this.clock.tick();
-
-            const appointmentForm = this.instance._appointmentPopup._appointmentForm;
-
-            assert.deepEqual(appointmentForm.option('formData').startDate, new Date(2015, 1, 9, 13), 'Form data is correct');
-        } finally {
-            tzOffsetStub.restore();
-        }
-    });
-
-    QUnit.test('Appointment should not be updated if \'cancel\' flag is defined as true during async operation', function(assert) {
-        const appointments = [{ startDate: new Date(2015, 1, 9, 16), endDate: new Date(2015, 1, 9, 17), text: 'caption' }];
-        const dataSource = new DataSource({
-            store: appointments
-        });
-
-        this.createInstance({
-            onAppointmentUpdating: function(args) {
-                args.cancel = $.Deferred();
-                setTimeout(function() {
-                    args.cancel.resolve(true);
-                }, 200);
-            },
-            dataSource: dataSource,
-            currentDate: new Date(2015, 1, 9)
-        });
-
-        this.instance.updateAppointment(appointments[0], { startDate: new Date(), text: 'Appointment 1' });
-        this.clock.tick(200);
-
-        assert.deepEqual(dataSource.items(), [{ startDate: new Date(2015, 1, 9, 16), endDate: new Date(2015, 1, 9, 17), text: 'caption' }], 'Update operation is canceled');
-    });
-
-    QUnit.test('Appointment should be returned to the initial state if \'cancel\' flag is defined as true during async operation', function(assert) {
-        this.createInstance({
-            onAppointmentUpdating: function(args) {
-                const d = $.Deferred();
-                args.cancel = d.promise();
-                setTimeout(function() {
-                    d.reject();
-                }, 200);
-            },
-            currentView: 'week',
-            dataSource: [{ startDate: new Date(2015, 1, 11), endDate: new Date(2015, 1, 13), text: 'caption' }],
-            firstDayOfWeek: 1,
-            currentDate: new Date(2015, 1, 9)
-        });
-
-        const $appointment = $(this.instance.$element().find('.dx-scheduler-appointment').eq(0));
-        const initialLeftPosition = translator.locate($appointment).left;
-        const cellWidth = this.instance.$element().find('.dx-scheduler-all-day-table-cell').eq(0).outerWidth();
-        const pointer = pointerMock(this.instance.$element().find('.dx-resizable-handle-left').eq(0)).start();
-
-        pointer.dragStart().drag(-cellWidth * 2, 0).dragEnd();
-        this.clock.tick(200);
-        assert.equal(translator.locate(this.instance.$element().find('.dx-scheduler-appointment').eq(0)).left, initialLeftPosition, 'Left position is OK');
-    });
-
-    QUnit.test('Appointment should have initial position if \'cancel\' flag is defined as true during update operation', function(assert) {
-        const appointments = [{ startDate: new Date(2015, 1, 9, 1), endDate: new Date(2015, 1, 9, 2), text: 'caption' }];
-        const dataSource = new DataSource({
-            store: appointments
-        });
-
-        this.createInstance({
-            onAppointmentUpdating: function(args) {
-                args.cancel = true;
-            },
-            dataSource: dataSource,
-            firstDayOfWeek: 1,
-            currentDate: new Date(2015, 1, 9)
-        });
-
-        let $appointment = $(this.instance.$element().find('.dx-scheduler-appointment').eq(0));
-        const initialPosition = translator.locate($appointment);
-
-        $(this.instance.$element().find('.dx-scheduler-date-table-cell').eq(5)).trigger(dragEvents.enter);
-
-        pointerMock($appointment)
-            .start()
-            .down(initialPosition.left + 10, initialPosition.top + 10)
-            .move(initialPosition.left + 10, initialPosition.top + 100)
-            .up();
-
-        $appointment = $(this.instance.$element().find('.dx-scheduler-appointment').eq(0));
-        assert.deepEqual(translator.locate($appointment), initialPosition, 'Appointments position is OK');
-    });
-
-    QUnit.test('Appointment should have initial size if \'cancel\' flag is defined as true during update operation (day view)', function(assert) {
-        this.createInstance({
-            onAppointmentUpdating: function(args) {
-                args.cancel = true;
-            },
-            dataSource: [{ startDate: new Date(2015, 1, 9, 1), endDate: new Date(2015, 1, 9, 2), text: 'caption' }],
-            firstDayOfWeek: 1,
-            currentDate: new Date(2015, 1, 9)
-        });
-
-        const $appointment = $(this.instance.$element().find('.dx-scheduler-appointment').eq(0));
-        const initialHeight = $appointment.outerHeight();
-        const cellHeight = this.instance.$element().find('.dx-scheduler-date-table-cell').eq(0).outerHeight();
-
-        const pointer = pointerMock(this.instance.$element().find('.dx-resizable-handle-bottom').eq(0)).start();
-        pointer.dragStart().drag(0, cellHeight * 2).dragEnd();
-
-        assert.equal(this.instance.$element().find('.dx-scheduler-appointment').eq(0).outerHeight(), initialHeight, 'Height is OK');
-    });
-
-    QUnit.test('Appointment should have initial size if "cancel" flag is defined as true during update operation (month view)', function(assert) {
-        this.createInstance({
-            onAppointmentUpdating: function(args) {
-                args.cancel = true;
-            },
-            views: ['month'],
-            currentView: 'month',
-            dataSource: [{ startDate: new Date(2015, 1, 9, 1), endDate: new Date(2015, 1, 9, 2), text: 'caption' }],
-            firstDayOfWeek: 1,
-            currentDate: new Date(2015, 1, 9)
-        });
-
-        const $appointment = $(this.instance.$element().find('.dx-scheduler-appointment').eq(0));
-        const initialWidth = $appointment.outerWidth();
-        const cellWidth = this.instance.$element().find('.dx-scheduler-date-table-cell').eq(0).outerWidth();
-
-        const pointer = pointerMock(this.instance.$element().find('.dx-resizable-handle-right').eq(0)).start();
-        pointer.dragStart().drag(cellWidth * 2, 0).dragEnd();
-
-        assert.roughEqual(this.instance.$element().find('.dx-scheduler-appointment').eq(0).outerWidth(), initialWidth, 0.5, 'Width is OK');
-    });
-
-    QUnit.test('Appointment should have initial size if \'cancel\' flag is defined as true during update operation (all day)', function(assert) {
-        this.createInstance({
-            onAppointmentUpdating: function(args) {
-                args.cancel = true;
-            },
-            currentView: 'week',
-            dataSource: [{ startDate: new Date(2015, 1, 9), endDate: new Date(2015, 1, 10), text: 'caption', allDay: true }],
-            firstDayOfWeek: 1,
-            currentDate: new Date(2015, 1, 9)
-        });
-
-        const $appointment = $(this.instance.$element().find('.dx-scheduler-appointment').eq(0));
-        const initialWidth = $appointment.outerWidth();
-        const cellWidth = this.instance.$element().find('.dx-scheduler-all-day-table-cell').eq(0).outerWidth();
-
-        const pointer = pointerMock(this.instance.$element().find('.dx-resizable-handle-right').eq(0)).start();
-        pointer.dragStart().drag(cellWidth * 2, 0).dragEnd();
-
-        assert.roughEqual(this.instance.$element().find('.dx-scheduler-appointment').eq(0).outerWidth(), initialWidth, 1, 'Width is OK');
-    });
-
-    QUnit.test('Appointment should have initial size if \'cancel\' flag is defined as true during update operation (if appointment takes few days)', function(assert) {
-        this.createInstance({
-            onAppointmentUpdating: function(args) {
-                args.cancel = true;
-            },
-            currentView: 'week',
-            dataSource: [{ startDate: new Date(2015, 1, 9), endDate: new Date(2015, 1, 11), text: 'caption' }],
-            firstDayOfWeek: 1,
-            currentDate: new Date(2015, 1, 9)
-        });
-
-        const $appointment = $(this.instance.$element().find('.dx-scheduler-appointment').eq(0));
-        const initialWidth = $appointment.outerWidth();
-        const cellWidth = this.instance.$element().find('.dx-scheduler-all-day-table-cell').eq(0).outerWidth();
-
-        const pointer = pointerMock(this.instance.$element().find('.dx-resizable-handle-right').eq(0)).start();
-        pointer.dragStart().drag(cellWidth * 3, 0).dragEnd();
-
-        assert.roughEqual(this.instance.$element().find('.dx-scheduler-appointment').eq(0).outerWidth(), 1.1, initialWidth, 'Width is OK');
-    });
-
-    QUnit.test('Appointment should have initial left coordinate if \'cancel\' flag is defined as true during resize operation', function(assert) {
-        this.createInstance({
-            onAppointmentUpdating: function(args) {
-                args.cancel = true;
-            },
-            currentView: 'week',
-            dataSource: [{ startDate: new Date(2015, 1, 11), endDate: new Date(2015, 1, 13), text: 'caption' }],
-            firstDayOfWeek: 1,
-            currentDate: new Date(2015, 1, 9)
-        });
-
-        const $appointment = $(this.instance.$element().find('.dx-scheduler-appointment').eq(0));
-        const initialLeftPosition = translator.locate($appointment).left;
-        const cellWidth = this.instance.$element().find('.dx-scheduler-all-day-table-cell').eq(0).outerWidth();
-        const pointer = pointerMock(this.instance.$element().find('.dx-resizable-handle-left').eq(0)).start();
-
-        pointer.dragStart().drag(-cellWidth * 2, 0).dragEnd();
-
-        assert.equal(translator.locate(this.instance.$element().find('.dx-scheduler-appointment').eq(0)).left, initialLeftPosition, 'Left position is OK');
-    });
-
-    QUnit.test('Appointment should have initial top coordinate if \'cancel\' flag is defined as true during resize operation', function(assert) {
-        this.createInstance({
-            onAppointmentUpdating: function(args) {
-                args.cancel = true;
-            },
-            currentView: 'week',
-            dataSource: [{ startDate: 1423620000000, endDate: 1423627200000, text: 'caption' }],
-            firstDayOfWeek: 1,
-            currentDate: new Date(2015, 1, 9)
-        });
-
-        const $appointment = $(this.instance.$element().find('.dx-scheduler-appointment').eq(0));
-        const initialTopPosition = translator.locate($appointment).top;
-        const cellHeight = this.instance.$element().find('.dx-scheduler-all-day-table-cell').eq(0).outerHeight();
-        const pointer = pointerMock(this.instance.$element().find('.dx-resizable-handle-top').eq(0)).start();
-
-        pointer.dragStart().drag(0, -cellHeight * 2).dragEnd();
-
-        assert.equal(translator.locate(this.instance.$element().find('.dx-scheduler-appointment').eq(0)).top, initialTopPosition, 'Top position is OK');
-    });
-
-    QUnit.test('onAppointmentUpdated', function(assert) {
-        const updatedSpy = sinon.spy(noop);
-        const oldData = [{ startDate: new Date(2015, 1, 9, 16), endDate: new Date(2015, 1, 9, 17), text: 'caption' }];
-        const newData = { startDate: new Date(2015, 1, 10, 16), endDate: new Date(2015, 1, 10, 17), text: 'title' };
-
-        this.createInstance({
-            onAppointmentUpdated: updatedSpy,
-            dataSource: new DataSource({ store: oldData }),
-            currentDate: new Date(2015, 1, 9)
-        });
-
-        this.instance.updateAppointment(oldData[0], newData);
-        this.clock.tick();
-
-        const args = updatedSpy.getCall(0).args[0];
-
-        assert.ok(updatedSpy.calledOnce, 'onAppointmentUpdated was called');
-        assert.equal(args.element, this.instance.element(), 'Element field is OK');
-        assert.equal(args.component, this.instance, 'Component field is OK');
-        assert.deepEqual(args.appointmentData, newData, 'newData field is OK');
-        assert.strictEqual(args.error, undefined, 'Error field is not defined');
-    });
-
-    QUnit.test('onAppointmentUpdated should have error field in args if an error occurs while data updating', function(assert) {
-        const updatedSpy = sinon.spy(noop);
-        const oldData = [{ startDate: new Date(2015, 1, 9, 16), endDate: new Date(2015, 1, 9, 17), text: 'caption' }];
-        const newData = { startDate: new Date(2015, 1, 10, 16), endDate: new Date(2015, 1, 10, 17), text: 'title' };
-
-        this.createInstance({
-            onAppointmentUpdated: updatedSpy,
-            dataSource: new DataSource({
-                store: new CustomStore({
-                    load: function(options) {
-                        const d = $.Deferred();
-                        d.resolve(oldData);
-                        return d.promise();
-                    },
-                    update: function() {
-                        return $.Deferred().reject(new Error('Unknown error occurred'));
-                    }
-                })
-            })
-        });
-
-        this.instance.updateAppointment(oldData[0], newData);
-        this.clock.tick();
-
-        const error = updatedSpy.getCall(0).args[0].error;
-
-        assert.ok(error instanceof Error, 'Error field is defined');
-        assert.equal(error.message, 'Unknown error occurred', 'Error message is OK');
-    });
-
-    QUnit.test('onAppointmentDeleting', function(assert) {
-        const deletingSpy = sinon.spy(noop);
-        const appointments = [
-            { startDate: new Date(2015, 3, 29, 5), text: 'Appointment 1', endDate: new Date(2015, 3, 29, 6) }
-        ];
-
-        this.createInstance({
-            onAppointmentDeleting: deletingSpy,
-            currentDate: new Date(2015, 3, 29),
-            dataSource: new DataSource({
-                store: appointments
-            })
-        });
-
-        this.instance.deleteAppointment(appointments[0]);
-        this.clock.tick();
-
-        const args = deletingSpy.getCall(0).args[0];
-
-        assert.ok(deletingSpy.calledOnce, 'onAppointmentDeleting was called');
-        assert.equal(args.element, this.instance.element(), 'Element field is OK');
-        assert.equal(args.component, this.instance, 'Component field is OK');
-        assert.deepEqual(args.appointmentData, { startDate: new Date(2015, 3, 29, 5), text: 'Appointment 1', endDate: new Date(2015, 3, 29, 6) }, 'Appointment field is OK');
-        assert.strictEqual(args.cancel, false, '\'Cancel\' flag is OK');
-    });
-
-    QUnit.test('Appointment should not be deleted if \'cancel\' flag is defined as true', function(assert) {
-        const appointments = [{ startDate: new Date(2015, 1, 9, 16), endDate: new Date(2015, 1, 9, 17), text: 'caption' }];
-        const dataSource = new DataSource({
-            store: appointments
-        });
-
-        this.createInstance({
-            onAppointmentDeleting: function(args) {
-                args.cancel = true;
-            },
-            dataSource: dataSource,
-            currentDate: new Date(2015, 1, 9)
-        });
-
-        this.instance.deleteAppointment(appointments[0]);
-        this.clock.tick();
-
-        assert.deepEqual(dataSource.items(), [{ startDate: new Date(2015, 1, 9, 16), endDate: new Date(2015, 1, 9, 17), text: 'caption' }], 'Delete operation is canceled');
-    });
-
-    QUnit.test('Appointment should not be deleted if \'cancel\' flag is defined as true during async operation', function(assert) {
-        const appointments = [{ startDate: new Date(2015, 1, 9, 16), endDate: new Date(2015, 1, 9, 17), text: 'caption' }];
-        const dataSource = new DataSource({
-            store: appointments
-        });
-
-        this.createInstance({
-            onAppointmentDeleting: function(args) {
-                args.cancel = $.Deferred();
-                setTimeout(function() {
-                    args.cancel.resolve(true);
-                }, 200);
-            },
-            dataSource: dataSource,
-            currentDate: new Date(2015, 1, 9)
-        });
-
-        this.instance.deleteAppointment(appointments[0]);
-        this.clock.tick(200);
-
-        assert.deepEqual(dataSource.items(), [{ startDate: new Date(2015, 1, 9, 16), endDate: new Date(2015, 1, 9, 17), text: 'caption' }], 'Delete operation is canceled');
-    });
-
-    QUnit.test('Appointment should be deleted correctly if \'cancel\' flag is defined as false during async operation', function(assert) {
-        const appointments = [{ startDate: new Date(2015, 1, 9, 16), endDate: new Date(2015, 1, 9, 17), text: 'caption' }];
-        const dataSource = new DataSource({
-            store: appointments
-        });
-
-        this.createInstance({
-            onAppointmentDeleting: function(args) {
-                args.cancel = $.Deferred();
-                setTimeout(function() {
-                    args.cancel.resolve(false);
-                }, 200);
-            },
-            dataSource: dataSource,
-            currentDate: new Date(2015, 1, 9)
-        });
-
-        this.instance.deleteAppointment(appointments[0]);
-        this.clock.tick(200);
-
-        assert.equal(dataSource.items().length, 0, 'Delete operation is completed');
-    });
-
-    QUnit.test('onAppointmentDeleted', function(assert) {
-        const deletedSpy = sinon.spy(noop);
-        const appointments = [
-            { startDate: new Date(2015, 3, 29, 5), text: 'Appointment 1', endDate: new Date(2015, 3, 29, 6) }
-        ];
-
-        this.createInstance({
-            onAppointmentDeleted: deletedSpy,
-            currentDate: new Date(2015, 3, 29),
-            dataSource: new DataSource({
-                store: appointments
-            })
-        });
-
-        this.instance.deleteAppointment(appointments[0]);
-        this.clock.tick();
-
-        const args = deletedSpy.getCall(0).args[0];
-        assert.ok(deletedSpy.calledOnce, 'onAppointmentDeleted was called');
-        assert.equal(args.element, this.instance.element(), 'Element field is OK');
-        assert.equal(args.component, this.instance, 'Component field is OK');
-        assert.deepEqual(args.appointmentData, { startDate: new Date(2015, 3, 29, 5), text: 'Appointment 1', endDate: new Date(2015, 3, 29, 6) }, 'newData field is OK');
-        assert.strictEqual(args.error, undefined, 'Error field is not defined');
-    });
-
-    QUnit.test('onAppointmentDeleted should have error field in args if an error occurs while data deleting', function(assert) {
-        const deletedSpy = sinon.spy(noop);
-
-        this.createInstance({
-            onAppointmentDeleted: deletedSpy,
-            dataSource: new DataSource({
-                store: new CustomStore({
-                    load: noop,
-                    remove: function() {
-                        return $.Deferred().reject(new Error('Unknown error occurred'));
-                    }
-                })
-            })
-        });
-
-        this.instance.deleteAppointment({});
-        this.clock.tick();
-
-        const error = deletedSpy.getCall(0).args[0].error;
-
-        assert.ok(error instanceof Error, 'Error field is defined');
-        assert.equal(error.message, 'Unknown error occurred', 'Error message is OK');
     });
 
     QUnit.test('onAppointmentRendered', function(assert) {
