+++ testing/tests/DevExpress.ui.widgets.scheduler/integration.recurringAppointments.tests.js
@@ -9,14 +9,13 @@ import { DataSource } from 'data/data_source/data_source';
 import dateSerialization from 'core/utils/date_serialization';
 import { createWrapper, SchedulerTestWrapper, isDesktopEnvironment } from '../../helpers/scheduler/helpers.js';
 import dateUtils from 'core/utils/date';
-import ArrayStore from 'data/array_store';
 import timeZoneUtils from 'ui/scheduler/utils.timeZone';
 
 import 'common.css!';
 import 'generic_light.css!';
 import 'ui/scheduler/ui.scheduler';
 
-const { module, test } = QUnit;
+const { module } = QUnit;
 const toMs = dateUtils.dateToMilliseconds;
 
 QUnit.testStart(function() {
@@ -60,53 +59,6 @@ QUnit.testStart(function() {
             this.clock.restore();
         }
     }, function() {
-        if(isDesktopEnvironment()) {
-            test('Excluded appointment\'s key property shouldn\'t equal to parent appointment(T929772)', function(assert) {
-                const data = [{
-                    id: 1,
-                    text: 'Appointment',
-                    startDate: new Date(2017, 4, 22, 1, 30),
-                    endDate: new Date(2017, 4, 22, 2, 30),
-                    recurrenceRule: 'FREQ=DAILY',
-                }];
-
-                const scheduler = createWrapper({
-                    dataSource: {
-                        store: new ArrayStore({
-                            data: data,
-                            key: 'id'
-                        })
-                    },
-                    views: ['week'],
-                    currentView: 'week',
-                    recurrenceEditMode: 'occurrence',
-                    currentDate: new Date(2017, 4, 22),
-                    onAppointmentAdding: e => {
-                        assert.equal(e.appointmentData.id, undefined, 'key property \'id\' shouldn\'t exist in appointment on onAppointmentAdding event');
-                    },
-                    onAppointmentAdded: e => {
-                        assert.equal(e.appointmentData.id.length, 36, 'key property \'id\' should be equal GUID in appointment on onAppointmentAdded event');
-                        assert.notStrictEqual(data[0].id, e.appointmentData.id, 'Excluded appointment\'s key property shouldn\'t equal to parent appointment');
-                    },
-                    height: 600
-                });
-
-                scheduler.appointmentList[3].drag.toCell(39);
-
-                const appointments = scheduler.instance.getDataSource().items();
-                const recurrenceAppointment = appointments[0];
-                const excludedAppointment = appointments[1];
-
-                const expectedDate = new Date(excludedAppointment.startDate);
-                expectedDate.setHours(recurrenceAppointment.startDate.getHours() + 1);
-
-                assert.equal(excludedAppointment.startDate.valueOf(), expectedDate.valueOf(), 'appointment should be shifted down');
-                assert.equal(excludedAppointment.id.length, 36, 'id property should be equal GUID');
-
-                assert.expect(5);
-            });
-        }
-
         QUnit.test('Tasks should be duplicated according to recurrence rule', function(assert) {
             const tasks = [
                 { text: 'One', startDate: new Date(2015, 2, 16), endDate: new Date(2015, 2, 16, 2), recurrenceRule: 'FREQ=DAILY;INTERVAL=4' },
